@@ -12,45 +12,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@ngx-translate/core": "^9.0.2",
+    "@angular/animations": "^6.1.0",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "@ngx-translate/core": "^10.0.2",
     "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
-    "ng-http-loader": "^0.6.0",
-    "ngx-bootstrap": "^2.0.0-beta.8",
-    "ngx-moment": "^2.0.0-rc.0",
-    "ngx-pagination": "^3.1.0",
-    "ngx-toastr": "^8.0.0",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14"
+    "core-js": "^2.5.4",
+    "moment": "^2.22.2",
+    "ng-http-loader": "^3.1.2",
+    "ngx-bootstrap": "^3.0.1",
+    "ngx-moment": "^3.0.1",
+    "ngx-pagination": "^3.2.1",
+    "ngx-toastr": "^9.1.0",
+    "rxjs": "~6.2.0",
+    "rxjs-compat": "^6.3.3",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular/cli": "^6.2.4",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.3.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "^3.0.0",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "^5.4.1",
     "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "tslint": "~5.11.0",
+    "typescript": "~2.9.2"
   }
 }
